Replace GuiUtils to PanelUtils class.
Add PanelUtils dependency to the MCG.
