+++ src/main/java/world/bentobox/magiccobblestonegenerator/panels/utils/SelectBundlePanel.java
@@ -8,14 +8,15 @@ import java.util.stream.Collectors;
 
 import org.bukkit.Material;
 
+import lv.id.bonne.panelutils.PanelUtils;
 import world.bentobox.bentobox.api.panels.PanelItem;
 import world.bentobox.bentobox.api.panels.builders.PanelBuilder;
 import world.bentobox.bentobox.api.panels.builders.PanelItemBuilder;
 import world.bentobox.magiccobblestonegenerator.database.objects.GeneratorBundleObject;
 import world.bentobox.magiccobblestonegenerator.database.objects.GeneratorTierObject;
 import world.bentobox.magiccobblestonegenerator.panels.CommonPanel;
-import world.bentobox.magiccobblestonegenerator.panels.GuiUtils;
 import world.bentobox.magiccobblestonegenerator.utils.Constants;
+import world.bentobox.magiccobblestonegenerator.utils.Utils;
 
 
 /**
@@ -62,7 +63,7 @@ public class SelectBundlePanel extends CommonPanel
         PanelBuilder panelBuilder =
             new PanelBuilder().user(this.user).name(this.user.getTranslation(Constants.TITLE + "select-bundle"));
 
-        GuiUtils.fillBorder(panelBuilder, Material.BLUE_STAINED_GLASS_PANE);
+        PanelUtils.fillBorder(panelBuilder, Material.BLUE_STAINED_GLASS_PANE);
 
         // Maximal elements in page.
         final int MAX_ELEMENTS = 21;
@@ -147,7 +148,7 @@ public class SelectBundlePanel extends CommonPanel
                 break;
             }
             case PREVIOUS -> {
-                count = GuiUtils.getPreviousPage(this.pageIndex, this.maxPageIndex);
+                count = Utils.getPreviousPage(this.pageIndex, this.maxPageIndex);
                 description.add(this.user
                         .getTranslationOrNothing(Constants.BUTTON + button.name().toLowerCase() + ".description",
                                 Constants.NUMBER, String.valueOf(count)));
@@ -166,7 +167,7 @@ public class SelectBundlePanel extends CommonPanel
                 break;
             }
             case NEXT -> {
-                count = GuiUtils.getNextPage(this.pageIndex, this.maxPageIndex);
+                count = Utils.getNextPage(this.pageIndex, this.maxPageIndex);
                 description.add(this.user
                         .getTranslationOrNothing(Constants.BUTTON + button.name().toLowerCase() + ".description",
                                 Constants.NUMBER, String.valueOf(count)));
